requirements.txt
fastapi
uvicorn
gunicorn==20.1.0
aiofiles
pymongo
cloudinary
firebase-admin
SQLAlchemy
anyio
httpx
h11
python-dotenv
PyJWT
cryptography
passlib[bcrypt]
python-socketio
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
dnspython
grpcio
grpcio-status
h2
hpack
httpcore
hyperframe
idna
itsdangerous
Jinja2
MarkupSafe
msgpack
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
requests
rsa
six
sniffio
typing_extensions
urllib3
Werkzeug
python-multipart
gradio_client
numpy
scipy
soundfile


